fix(brain): Fix `githubMetrics` not being loaded (#135)
This regressed in https://github.com/getsentry/sentry-ci-tooling/pull/102 where we filter what modules to autoload based on module name === exported fn name, as there was a function I wanted to import only for testing.

I added a simple test, but this is a bit brittle and we should remove this in the future. e.g. instead of enforcing the names to be the same, we could simply move the exported fn to a different module.