ref(api): Remove repo install, cache org installation id
We can assume that our GH app is always installed org wide, so there is no need for a repo level installation.
Additionally, cache the installation ID in memory, we could even hard code it as this should never change, but we will just store it in memory for now.

There may be further optimizations to be had here by caching the Octokit client, but we can save that for later.