feat: Add button to quickly revert a failing commit (for testing only) (#210)
This adds a button to the `Required Checks` feature that slacks `#team-engineering` when a commit causes master to be red. The button allows the user to quickly revert the relevant commit (via https://github.com/getsentry/sentry-deploy-sync-hook/pull/34).

## Note
This commit is only to test at the moment. We will follow-up after some tests.

![image](https://user-images.githubusercontent.com/79684/122486151-55b20300-cf8d-11eb-8235-14e60ff8ef9f.png)