feat: Fix required checks to respect commit timestamp (#236)
This fixes required checks to correctly mark only older failing builds (based on commit timestamp) as fixed in Slack.

If we have two builds:
* `A` - which will pass CI and is merged before `B`
* `B` - which is fail CI and is merged after `A`

When `B` fails it will send a Slack message saying it is failing. Assuming `A` does not complete before `B` fails, when `A` finally finishes, it will *incorrectly* mark `B` as being fixed. When a build finishes, we need to compare the commits of 1) the build that finished with 2) the builds that failed and make sure that we only mark builds that have started after the build that finished.

Closes https://github.com/getsentry/eng-pipes/issues/234