feat(brain): Check for queued deploys after check run finishes (#246)
When a freight deploy is queued (all deploys start in the "queued" state), we save the list of commits in that deploy to a table. Then in the case where a CI run finishes *after* a deploy is queued, the Slack message the author receives will no longer have the "deploy" button as it's already been queued.