feat(utils): Add function to get latest deploy between two projects (#250)
Adds a utility function that will look for deploys between
getsentry/getsentry-frontend project and returns the deploy with the most
recent commit. This only works because they deploy from the same
repository.

E.g. getsentry last deployed commit `foo` and later on someone deploys
`getsentry-frontend@bar` (which is ahead of `foo`), the function should
return the deploy for `getsentry-frontend`.

Requires a migration to change the unique fields for the deploys table to
include `app_name`