feat(brain): Frontend-only deploy helper needs to check all commits in deploy
When considering a commit that has just passed CI checks... we consider to be available to be deployed via the frontend-only express deploy when *all* commits between this commit and the lastest deployed commit between the getsentry and getsentry-frontend commits, are frontend-only changes. Otherwise, you could be deploying a frontend change the depends on a backend change.