feat(brain): Save partial checkRun payload after required check
Save partial checkRun webhook payload after a required check fails. This is for the future so that we can retrieve and use it to make a better message after we resolve a failed build.

It currently wipes out the existing failed build info with the build info of the build that passes, so we lose some historical context in our Slack messages when a build resolves.