feat(brain): Save partial `checkRun` payload after required check fails (#257)
Save partial checkRun webhook payload after a required check fails. This is for the future so that we can retrieve and use it to make a better message after we resolve a failed build.

It currently wipes out the existing failed build info with the build info of the build that passes, so we lose some historical context in our Slack messages when a build resolves.

In the example below, `7678bad` was a commit after the failing commit that passed. The original message should not be referencing `7678bad`:

![image](https://user-images.githubusercontent.com/79684/141863692-bf33ad8c-a20a-4128-8f5a-410770d45b21.png)