feat(brain): Add restarting workflows on failed job steps (#279)
This will auto restart workflows if certain steps within a job fail. It also suppresses sending failure messages to Slack. The auto-restart will only ever happen once (e.g. only if the first run attempt fails, follow-up attempts will be treated as a failure).

The intention for this change is to reduce noise with our Slack notifications when builds fail due to intermittent CI issues. These issues are generally out of our control (e.g. when docker fails to pull an image, or GHA randomly chokes) and occur before we run any of our tests. This change should suppress a lot of the noise in `#team-engineering` so that when you do get pinged of a broken build, it should be due to a change our engineers made (that or their tests are flakey).