feat(migration): Add table for build failures (#285)
Adding a table to store build failures. Not using BigQuery for this data because it is not meant to be a metric but rather some debugging and logging to be able to access further down the road. This should be moved into BigQuery if the data does prove to be helpful.