ref: Use a bot access token to check for Github org membership
The Github API for checking organization membership will only show membership for people with the "private" setting if the API call is made by a member of that organization.

Since the Github application is not a user thus not part of the Sentry org, we need to use a bot account's Personal Access Token to correctly check that a user is part of the Github org.