Fix numbering bug in #409
The bug appeared because numbers were assigned to the output text before
sorting occurred. In addition to some DRY cleanup, we now create a
`SlackMessageUnorderedIssueItem` which stores all of the output strings
we'll need without numbering. Once we sort, we then map over the
now-sorted list to map all of our `SlackMessageUnorderedIssueItem`s to
`SlackMessageOrderedIssueItem`s via the new
`addOrderingToSlackMessageItem` helper function.